pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "web-monitor"
version = "0.1.0"
description = "Simple web monitoring service with email notifications"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "pydantic>=2.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.4",
]

[project.scripts]
web-monitor = "web_monitor.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
